backend/Avatar_LLM_Endpoint/requirements.txt
Flask
python-dotenv
flask-cors
langchain-openai
langchain-core
gunicorn
